replace model MaterialTemplate in material-state
