+++ libs/material/src/lib/template/components/template-material-form/template-material-form.component.ts
@@ -3,7 +3,7 @@ import { ControlContainer } from '@angular/forms';
 import { Currencies } from '../../../delivery/+state';
 import { MaterialControl } from '../../forms/material.form';
 import { FormElement } from '@blockframes/utils';
-import { MaterialTemplate } from '../../material/+state';
+import { MaterialTemplate } from '../../../material/+state';
 
 @Component({
   selector: '[formGroup] template-material-form',
