--- libs/material/src/lib/template/material/+state/index.ts
@@ -1,1 +1,0 @@
-export * from './template-material.model';
